refactor: 테스트 코드를 하위의 test 폴더로 이동
